Move tests to new class to avoid parameterized tests
