+++ rskj-core/src/test/java/org/ethereum/datasource/KeyValueDataSourceTest.java
@@ -8,8 +8,6 @@ import org.junit.rules.TemporaryFolder;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
-import java.io.File;
-import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -130,19 +128,6 @@ public class KeyValueDataSourceTest {
         updatedValues.put(keyToNull, null);
 
         keyValueDataSource.updateBatch(updatedValues, Collections.emptySet());
-    }
-
-    @Test
-    public void testShouldValidateKindWithEmptyDbDirAndResetDbFalseSuccessfully() {
-        KeyValueDataSource.validateDbKind(DbKind.ROCKS_DB, tempFolder.getRoot().getPath(), false);
-    }
-
-    @Test(expected = IllegalStateException.class)
-    public void testShouldThrowErrorWhenValidatingDifferentKinds() throws IOException {
-        FileWriter fileWriter = new FileWriter(new File(tempFolder.getRoot(), KeyValueDataSource.DB_KIND_PROPERTIES_FILE));
-        fileWriter.write("keyvalue.datasource=ROCKS_DB\n");
-        fileWriter.close();
-        KeyValueDataSource.validateDbKind(DbKind.LEVEL_DB, tempFolder.getRoot().getPath(), false);
     }
 
     private Map<ByteArrayWrapper, byte[]> generateRandomValuesToUpdate(int maxValuesToCreate) {
